+++ pkg/meshdiscovery/mesh/linkerd/linkerd.go
@@ -33,7 +33,7 @@ func NewLinkerdDiscoverySyncer() *linkerdDiscoverySyncer {
 }
 
 func (s *linkerdDiscoverySyncer) DiscoverMeshes(ctx context.Context, snap *v1.DiscoverySnapshot) (v1.MeshList, error) {
-	ctx = contextutils.WithLogger(ctx, fmt.Sprintf("istio-translation-sync-%v", snap.Hash()))
+	ctx = contextutils.WithLogger(ctx, fmt.Sprintf("linkerd-mesh-discovery-sync-%v", snap.Hash()))
 	logger := contextutils.LoggerFrom(ctx)
 
 	pods := snap.Pods.List()
