+++ pkg/meshdiscovery/mesh/syncer.go
@@ -1,4 +1,4 @@
-package discovery
+package mesh
 
 import (
 	"context"
